Make JSON goal control independent of the Octahedral Complex
