+++ src/dev/btietz/TC_goal/CMakeLists.txt
@@ -18,6 +18,7 @@ add_library(GoalSpine SHARED
                 BaseSpineModelGoal.cpp
                 FlemonsSpineModelGoal.cpp
                 SpineGoalControl.cpp
+                JSONGoalControl.cpp
                 )
                 
                 
@@ -25,5 +26,9 @@ add_executable(AppGoalTerrain
     BaseSpineModelGoal.cpp
     FlemonsSpineModelGoal.cpp
     SpineGoalControl.cpp
+    JSONGoalControl.cpp
     AppGoalTerrain.cpp
 )
+
+target_link_libraries(GoalSpine ${ENV_LIB_DIR}/libjsoncpp.a FileHelpers boost_program_options JSONControl)
+target_link_libraries(AppGoalTerrain ${ENV_LIB_DIR}/libjsoncpp.a FileHelpers boost_program_options JSONControl)
