+++ src/dev/btietz/TC_goal/JSONGoalControl.h
@@ -36,7 +36,6 @@ class neuralNetwork;
 class tgSpringCableActuator;
 class BaseSpineModelGoal;
 class tgSCASineControl;
-class OctahedralComplex;
 
 /**
  * JSONGoalControl learns the parameters for a CPG system on a
@@ -114,12 +113,8 @@ protected:
 
     virtual void setupCPGs(BaseSpineModelLearning& subject, array_2D nodeActions, array_4D edgeActions);
     
-    virtual void setupSaddleControllers(const OctahedralComplex* subject);
-    
     virtual array_2D scaleNodeActions (Json::Value actions);
 
-    void setGoalTensions(const BaseSpineModelGoal* subject);
-    
     std::vector<double> getGoalFeedback(const BaseSpineModelGoal* subject);
     
     std::vector<double> getFeedback(BaseSpineModelLearning& subject);
@@ -136,8 +131,6 @@ protected:
     /// @todo generalize this if we need more than one
     neuralNetwork* nn;
     
-    // A list of impedance controllers not associated with the CPG
-    std::vector<tgSCASineControl*> m_saddleControllers;
     
 };
